Pass user can view past orders test
- Commented out mailer in orders controller
- Consolidated test setup with login_user and create_user_order
